meal_backend_v2/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
sqlmodel==0.0.22
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
rapidfuzz==3.10.1
slowapi==0.1.9
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.20.0
pydantic-settings==2.5.2
pydantic==2.11.0
psycopg2-binary==2.9.9
email-validator==2.2.0
